Improve intendation handling of source code
Models may change the intendation of the code sent for
transformation or may remove al leading white space. The plugin
needs to normalize this for proper insertion and replaceing.

When code is selected for transformation the smallest indentation
is captured and stored in the selection info. Code that is sent to
and received from to the mode is dedented and when inserted the
source is indended again.